New sidebar
Created the new sidebar based off of the vizhub v2 one. Everything works except when you hover over for the edit and delete they show up for each file instead of just the one that your are hovering over. This probably has a simple fix that I am overlooking but I will try to fix it. The rename works but the delete button still needs functionality.