Change background color set up
Everything to change the theme is in the settings component but Im not sure how to update the theme on change. I have the state and that value going into the codemirror component but its not changing.

You have a much better understanding of the library so I was wondering if you knew how?