Attempt to Display Remote Selection
I tried to use the from and to as the selection range, which works except instead of highlighting the selection, it shows how the text would look when the selection is deleted. It might have something to do with CSS, but will have to look deeper into what's causing it.