Solved the hidden text issue
I looked into Decoration.mark and used that to change the background of the remote selection  rather than implement it in PresenceWidget. This solved the hidden/deleted text issue. A future improvement is making the selection color lighter so that the cursor colors are visible. I also need to test whether a multiple-lines selection works.