Lightened Selection and Attempted Including Cursor
I changed the color of the remote selection to be a lighter shade of yellow so that the cursor colors could be visible. However, I'm not sure if the remote cursor should be shown when there is a selection. There are a few options: 1) Not show the cursor and only the selection 2) Show the cursor only at the end (might need to take into account the selection direction because the end could be on the left or right) 3) Show a yellow left border and black right border around the selection. For both option 2 and 3, I solved an issue where adding a border using border-left or border-right shifts the text over by using box-shadow instead. However, when selecting multiple lines, every line shows the border color rather than only at the first and last lines.