Add popup notification when remote user leaves session
A popup now appears when the remote user leaves the session and disappears after 10 seconds. Things that could be improved is to set a standard size for the popup rather than have it expand depending on the text length, add a close button on the side, and add the username. I also need to figure out where to call the popup when the user joins the session.