Attempt to Prevent Notifications When Cursor Moves
When the remote cursor moves, a notification pops up every time saying that the user joined. A potential solution that I tried to implement is to create an array that keeps track of the users/presence ids that joined the session and then use a conditional statement that checks if the user already joined before displaying a notification. When the user leaves the session, the user would be removed from the array.