added rainbow brackets to the code editor
I used the repository linked in the issue: https://github.com/eriknewland/rainbowbrackets.git. The developer has not released a package so I copied over the javascript file and converted it to a typescript file to implement the feature. I also changed the colors that the developer had implemented in favor of higher contrast colors.